Don't run kbuildsycoca4 twice for an upgrade/downgrade/reinstall.
In exheres-0, pkg_postinst() is always run last for an
upgrade/downgrade/reinstall. So in pkg_postrm() we check whether
pkg_postinst() will run (it does for a new install, upgrade, downgrade,
reinstall, but not for an uninstall) and we only run kbuildsycoca4 if
it won't run in pkg_postinst().