Qt 4.6.1: Use phonon from kdesupport
KDE 4.4 needs a phonon newer than what Qt provides. Since Phonon won't
be synced for any Qt 4.6.x release, we won't be patching it any longer.
It's a pain to maintain & just plain ugly.

Thus, we build phonon in Qt, if the option is enabled, but don't install
it. Instead we post-depend on media-libs/phonon, which installs a newer
phonon and its backends. We have to build phonon in order to have HTML5
video tag support in webkit, or phonon support for pure Qt applications.

Course of action suggested by Thiago Maceira, Nokia developer.