Add kdeedu 4.5 rc2.
Kalzium used an internal snapshot of avogadro for its 3dviewer. In 4.5 it now
requires us to package avogadro separately and uses the system libs provided by
that.

Rocs had a dependency on graphviz which apparently we've failed to notice. It
was removed in 4.5 though.

There are no cmake parameters to disable libgps or liblocation for Marble.