Fix kde-l10n scm.
autogen.sh can handle all localisations so no point in the for loop.

autogen.sh uses the basename of the dirs to set CURRENT_LANG which is used when
installing things. So you'd see dirs like /usr/share/locale/kde-l10n-scm-en_GB/
getting installed.

To fix this unpack to ${WORK}/${localisation}.

Change WORK so we can use globs without catching the scripts or the build dirs.