Remove blockers, change kdelibs minimum version.
blockers are now handled in the kde.org exlib, individual packages only
have to specifiy their own blockers if they are version-specific (like
kate).

kdelibs-4.7.x is the last branch of the kdelibs-4 series, kde 4.8 will
be able to use kdelibs-4.7, so stop requiring kdelibs[>=$PV] and instead
require a minimum version which is kdelibs[>=4.7.0] for kde 4.7.x and
kdelibs[>=4.6.0] for kde 4.6.x.