plasma-desktop: X11/XCB is required now
From upstream's commit message:
"For the future we will turn X11 in a optional dependency again. But it
will never be that a component won't be built if we are building without
X11. Instead there will be a required Wayland backend and this means all
X11 specific code needs to become platform aware.

So the way forward is:
1. X11 required
2. X11 required, Wayland optional
3. X11 required, Wayland required
4. Wayland required, X11 optional

REVIEW: 118694

[1] https://mail.kde.org/pipermail/plasma-devel/2014-June/032259.html"

Change-Id: Id9de640be11e1fc32534aaf2b1fe63b151358d06